NMS-9288: Topology on Map is not properly displayed
Cleaning classes. Moved all the not related to enlinkd stuff
to AbstractLinkdTopologyProvider
use static methods whenever possible
